As this is going to be the sister project of jGnuCashLib, we here copy a) the structure of this repostory, b) the make/build environment, c) the two files that are going to define the very core: the xsd file and the file that has defined it: test.xml (which also is going to be, as the name indicates, the data file for the test cases).
Similar to GnuCash, one might think that it's not very difficult to get
an already-defined XSD file for KMyMoney. Turns out: Nope, that doesn't
exist. We have to write it ourselves, and since we do not want to do
that manually (at least not completely), we have had it generate from
test test.xml file.